arch: arm: adi-daq2.dtsi: mark the AD9680 as JESD204 top device
This change marks the SPI AD9680 device as a JESD204 device in the
device-tree. It also marks this as a top-device, meaning that the framework
will aggregate & sync other devices in order to make this one work.
i.e. so that the JESD204 links initialize or re-initialize when some
configuration changes.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>