iio: adc: ad9680: register init, enable & disable link callbacks
This change moves the link init, disable & enable into framework ops.
The aim is to move parts of the driver into the framework into a manner
that is reviewable.

The link_init cb initializes the link configuration in a similar way that
is done in `ad9680_setup()`. This way, there should be the possibility to
initialize the link parameters via DT configuration (where the case).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>